policymaker/pyproject.toml
[tool.poetry]
name = "dragon-hunters-policymaker"
version = "0.0.0"
description = ""
authors = ["Futrime <futrime@outlook.com>"]
readme = "README.md"
packages = [{include = "policymaker"}]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.6"
python-dotenv = "^1.0.0"
jsonschema = "^4.19.1"
openai = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
